webapi/tct-webaudio-w3c-tests/webaudio/audiocontext_decodeAudioData_base.html

<!DOCTYPE html>
<!--
Copyright (c) 2014 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Tan, Shiyou <shiyoux.tan@intel.com>
        Cui, Jieqiong <jieqiongx.cui@intel.com>
-->

<meta charset="utf-8">
<title>WebAudio Test: audiocontext_decodeAudioData_base</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://webaudio.github.io/web-audio-api/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/webaudio.js"></script>
<div id="log"></div>
<script>
var t = async_test("Check if callback argument of AudioContext.decodeAudioData equals to AudioBuffer if AudioContext.decodeAudioData decode success", {timeout: 30000});
setup({timeout:30000});
var context = new AudioContext();
assert_true("decodeAudioData" in context, "decodeAudioData method exist AudioContext interface");
var request = new XMLHttpRequest();
request.open('GET', url, true);
request.responseType = 'arraybuffer';
request.onload = function() {
  context.decodeAudioData(request.response, onSuccess, onError);
}
request.send();
function onSuccess (buffer) {
  var audioBuffer = context.decodeAudioData(request.response, onDecodeSuccess, onDecodeError);
}
function onDecodeSuccess () {
  t.done();
}
function onDecodeError () {
  t.step(function() {
    assert_unreached("audio file data decoding error");
  });
  t.done();
}
function onError () {
  t.step(function() {
    assert_unreached("XMLHttpRequest request get audio file failed");
  });
  t.done();
}
</script>
